third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/XMLHttpRequestException-expected.txt

Tests the properties of XMLHttpRequest exceptions.

PASS e.toString() is "NetworkError: A network error occurred."
PASS Object.prototype.toString.call(e) is "[object DOMException]"
FAIL Object.prototype.toString.call(e.__proto__) is not "[object DOMExceptionPrototype]". Was "[object Object]"
PASS e.constructor.toString() is "function DOMException() { [native code] }"
PASS e.constructor is window.DOMException
PASS e.NETWORK_ERR is e.constructor.NETWORK_ERR
PASS e.NETWORK_ERR is 19
